Somewhere for the percussion to go
**A sixth of the collection was silent.** Counted over all 774,268 files:
14.6% of every note played is a General MIDI percussion instrument the
vocabulary had no voice for. Read in, filed, listed, drawn in the piano
roll -- and dropped on the way to the synthesiser. Tambourine alone is
2.24% of the corpus, congas 2.45%, maracas 1.67%. Four libraries are
more than half percussion and `Africa` is 77%, which played nothing at
all.

The input table read twenty-one of the standard's forty-seven
instruments. The other twenty-six are now voices: clap, tambourine,
cowbell, vibraslap, bongos, congas, timbales, agogo, cabasa, maracas,
whistles, guiros, claves, wood blocks, cuica, triangles. One per
instrument the standard names rather than a judgement about which are
alike enough to fold -- folding is how a vocabulary comes to disagree
with the thing it describes, and the standard is not ambiguous.

Measured against the collection: **67.8% of notes could be played
before, 90.5% now**. `Africa` and `Europe` go from nothing to all of it;
the 360,000-file GM pack from 13% to 77%, Superior Drummer from 54% to
95%. Twenty-eight of the forty-nine libraries improve.

**The mask had a ceiling at thirty-one.** Mutes travel as a bitmask and
JavaScript's `<<` is 32-bit, so with forty voices `1 << 32` wraps to 1
and silencing the thirty-second drum would have silenced the kick. It is
arithmetic on the page now, good for fifty-three, and `uint64_t` on the
wire. The roster test asks for bit 39 and checks it arrives whole.

Voices are appended, never inserted: a DAW remembers automation by
parameter index, so the fourteen that existed keep theirs.

**And the piano roll was reading every library with the wrong map.** It
looked notes up in the Roland table whatever they came from -- the table
the *shipped corpus* was played on -- so an imported pack's tambourine
was not drawn at all and the roll quietly disagreed with what the player
would sound. It reads with the library's own map now. The kit's fourteen
rows are still always drawn, because the most useful thing about a beat
is often that there is no ride in it; the percussion is drawn only where
it is played, since forty rows of mostly nothing is the wall that
arrangement exists to avoid.

Two tests changed rather than broke. "Every voice is reachable from the
corpus" is no longer the invariant -- the corpus was played on a TD-11,
which has no congas -- so it is now "reachable from some kit". And the
classifier test asserted that hand percussion comes back poorly
covered, which was the best that could be said while there was nowhere
to put it; it now asserts the truth, that such a pack is General MIDI
and all of it plays.

+++ native/core/include/jamin/Roster.h
@@ -102,7 +102,10 @@ public:
         /// And which drums another window has asked this one to silence, as a
         /// bit per voice. A mute is a thing somebody does to a *track*, and the
         /// window they do it from is whichever one happens to be open.
-        uint32_t wantMutes { 0 };
+        // Wide enough for every voice, which is forty and may grow. A
+        // 32-bit mask put the thirty-second drum's bit onto the kick's.
+        // @see src/core/drumKits.js DRUM_VOICES
+        uint64_t wantMutes { 0 };
         uint64_t wantMutesRevision { 0 };
     };
 
@@ -170,10 +173,10 @@ public:
         fourteen booleans -- but it still has to be a request, because a page
         cannot reach into another page and the processor that plays the notes is
         the one that has to stop playing them. */
-    void requestVoiceMutes (const std::string& id, uint32_t mask);
+    void requestVoiceMutes (const std::string& id, uint64_t mask);
 
     /** What this instance has been asked to silence since it last looked. */
-    bool takeVoiceMutesRequest (const Handle& slot, uint32_t& mask, uint64_t& seen) const;
+    bool takeVoiceMutesRequest (const Handle& slot, uint64_t& mask, uint64_t& seen) const;
 
     /** Solo is a question about everybody, so it has to be asked of everybody. */
     bool anySoloed() const;
